feat(dismiss): Add ability to dismiss notifications
Adds the ability to dismiss all notifications in the future. This doesn't prevent the dialog from opening, only prevents the daemon mode from showing notifications even if there are deprecated packages installed. The dismiss button becomes a show button, and opening the dialog manually and clicking that will turn notifications back on.