feat(pages): Hook up page-switching logic
Allow users to navigate forward and backwards between the flatpak and Apt pages. Because we gracefully fail if an application was already installed/removed, duplicate actions should not have any effect. We also change which buttons are visible based on whether or not there are packages selected in the list.

See https://github.com/pop-os/desktop/pull/50#issuecomment-643357029 and https://github.com/pop-os/desktop/pull/50#issuecomment-643385089