feat(package): add simple configuration migration
Adds simple configuration migration to the transition app. This requires that
the configuration be identical between two application versions (or the
application can migrate them), that both the old and new configuration are
stored in standardized locations and don't change based on other configuration,
that the old and new configuration are both stored in a directory located
somewhere within the user's home directory, and that copying the contents of
the old directory (and their attributes) to the new directory will preserve
the configiuration. Use cases outside this are currently out-of-scope and
should be handled manually by the user.

Also adds such configuration migration to Chromium.

https://github.com/pop-os/transition/pull/13#pullrequestreview-562990340