fix: add more granular control over the apt package system
Instead of blindly hoping the package system is unlocked (and only making atomic changes between each call), we should instead obtain a process-wide lock over dpkg, perform our changes, and then only release the lock once we're finished with the package system. This allows us to wait for a package manager lock from another process to be released, obtain the lock, operate on the system with it locked (preventing changes by other processes), and then release the lock so that other processes can operate on the cache. This also allows us to display more granular status to the user, which can help them understand what the system is waiting on (and allowing them to take action if needed to speed the process along).

Fixes #22